[Build query] Correctly handle phrases filter that specify only one single value
Contrary to the phrase filter type which always specify a meta.params.query, the phrases filter type never specify a meta.params.query even when a single value is selected.
This anomaly trigger an error generated by the MonoFieldQueryBase included in the elastic-builder dependency and prevent the generation of the report.

Signed-off-by: Kévin Masseix <masseix.kevin@gmail.com>